#3 refactor: Type 파일로 이동
